refactor(tina): rename pages folder to tina
