+++ apps/web/src/app/[locale]/(tina)/[slug]/page.tsx
