+++ apps/web/src/app/[locale]/(tina)/components/page-content.tsx
