+++ apps/web/src/app/cms/blocks/highlight.tsx
@@ -13,7 +13,7 @@ import { Section } from '@repo/design-system/components/section'
 import { useTranslations } from '@repo/i18n/client'
 import * as React from 'react'
 import { tinaField } from 'tinacms/dist/react'
-import { getPagePath } from '~/app/[locale]/(pages)/utils'
+import { getPagePath } from '~/app/[locale]/(tina)/utils'
 import { type PageBlocksHighlight } from '~/app/cms/types'
 
 const BlockLayout = ({ children }: { children: React.ReactNode }) => {
